# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api structure",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/parent/my-children": {
            "get": {
                "tags": [
                    "AuthParent"
                ],
                "summary": "myChildren",
                "description": "myChildren",
                "operationId": "authMyChildren",
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/parent/all-courses": {
            "get": {
                "tags": [
                    "AuthParent"
                ],
                "summary": "AllCourses",
                "description": "AllCourses",
                "operationId": "authParentAllCourses",
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/parent/my-cards": {
            "get": {
                "tags": [
                    "AuthParent"
                ],
                "summary": "MyCards",
                "description": "MyCards",
                "operationId": "authParentMyCards",
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/parent/add-card": {
            "post": {
                "tags": [
                    "AuthParent"
                ],
                "summary": "Add new card",
                "description": "Add Card",
                "operationId": "addCardAuthParent",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_number",
                                    "card_expiration",
                                    "card_token"
                                ],
                                "properties": {
                                    "card_number": {
                                        "type": "string",
                                        "example": "1234567890123456"
                                    },
                                    "card_expiration": {
                                        "type": "string",
                                        "example": "01/25"
                                    },
                                    "card_token": {
                                        "type": "string",
                                        "example": "4345678987653773hgfkdfu34hf3fhiuerifr4345678987653773hgfkdfu34hf3fhiuerifr4345678987653773hgfkdfu34hf3fhiuerifr"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/parent/delete-card/{id}": {
            "delete": {
                "tags": [
                    "AuthParent"
                ],
                "summary": "Delete card",
                "description": "Delete card",
                "operationId": "deleteCardAuthParent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID to fetch the targeted campaigns.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/parent/cashier": {
            "get": {
                "tags": [
                    "AuthParent"
                ],
                "summary": "Get Cashier",
                "description": "GetCashier",
                "operationId": "authParentGetCashier",
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/parent/pay-for-course": {
            "post": {
                "tags": [
                    "AuthParent"
                ],
                "summary": "Pay for course",
                "description": "Pay for course",
                "operationId": "payForCourseAuthParent",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "course_id",
                                    "card_id"
                                ],
                                "properties": {
                                    "student_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "course_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "card_id": {
                                        "type": "numeric",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}